move functions to lib
