+++ src/main.rs
@@ -1,169 +1,10 @@
 use crate::config::{BLACK, BLUE, CYAN, DARK_GRAY, GAP, GREEN, LIGHT_BLUE, LIGHT_GRAY, NC, PURPLE, RED, WHITE, YELLOW};
-use fortfetch::{get_cpu_model, get_desktop_environment, get_disk_info, get_gpu_model, get_package_count, Uptime};
+use fortfetch::{get_battery_info, get_cpu_model, get_cpu_usage, get_desktop_environment, get_disk_info, get_gpu_model, get_load_average, get_locale_info, get_network_info, get_package_count, get_processes_count, get_resolution, get_shell, get_temperature, get_terminal, get_users_count, Uptime};
 use std::process::Command;
 use std::fs;
 use std::env;
 
 mod config;
-
-fn get_shell() -> String {
-    std::env::var("SHELL")
-        .unwrap_or_else(|_| "?".to_string())
-        .split('/')
-        .last()
-        .unwrap_or("?")
-        .to_string()
-}
-
-fn get_terminal() -> String {
-    std::env::var("TERM")
-        .unwrap_or_else(|_| "?".to_string())
-}
-
-fn get_resolution() -> String {
-    if let Ok(output) = Command::new("xrandr").output() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        for line in stdout.lines() {
-            if line.contains("*") && line.contains("x") {
-                if let Some(res) = line.split_whitespace().find(|s| s.contains("x")) {
-                    return res.to_string();
-                }
-            }
-        }
-    }
-    
-    if let Ok(output) = Command::new("xdpyinfo").output() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        for line in stdout.lines() {
-            if line.trim().starts_with("dimensions:") {
-                if let Some(dims) = line.split(':').nth(1) {
-                    if let Some(res) = dims.trim().split_whitespace().next() {
-                        return res.to_string();
-                    }
-                }
-            }
-        }
-    }
-    
-    if env::var("WAYLAND_DISPLAY").is_ok() {
-        if let Ok(output) = Command::new("wlr-randr").output() {
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            for line in stdout.lines() {
-                if line.contains("current") && line.contains("x") {
-                    if let Some(res) = line.split_whitespace().find(|s| s.contains("x")) {
-                        return res.to_string();
-                    }
-                }
-            }
-        }
-    }
-    
-    "?".to_string()
-}
-
-fn get_load_average() -> String {
-    if let Ok(contents) = fs::read_to_string("/proc/loadavg") {
-        let parts: Vec<&str> = contents.split_whitespace().collect();
-        if parts.len() >= 3 {
-            return format!("{} {} {}", parts[0], parts[1], parts[2]);
-        }
-    }
-    "?".to_string()
-}
-
-fn get_users_count() -> String {
-    if let Ok(output) = Command::new("who").output() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        let count = stdout.lines().count();
-        return count.to_string();
-    }
-    "?".to_string()
-}
-
-fn get_battery_info() -> String {
-    if let Ok(entries) = fs::read_dir("/sys/class/power_supply") {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.file_name().unwrap_or_default().to_string_lossy().starts_with("BAT") {
-                if let Ok(capacity) = fs::read_to_string(path.join("capacity")) {
-                    if let Ok(status) = fs::read_to_string(path.join("status")) {
-                        return format!("{}% [{}]", capacity.trim(), status.trim());
-                    }
-                }
-            }
-        }
-    }
-    "Подключен к сети".to_string()
-}
-
-fn get_temperature() -> String {
-    if let Ok(temp_str) = fs::read_to_string("/sys/class/thermal/thermal_zone0/temp") {
-        if let Ok(temp) = temp_str.trim().parse::<i32>() {
-            return format!("{}°C", temp / 1000);
-        }
-    }
-    "?".to_string()
-}
-
-fn get_processes_count() -> String {
-    if let Ok(entries) = fs::read_dir("/proc") {
-        let count = entries
-            .flatten()
-            .filter(|entry| {
-                entry.file_name().to_string_lossy().chars().all(|c| c.is_ascii_digit())
-            })
-            .count();
-        return count.to_string();
-    }
-    "?".to_string()
-}
-
-fn get_cpu_usage() -> String {
-    if let Ok(contents) = fs::read_to_string("/proc/stat") {
-        if let Some(cpu_line) = contents.lines().next() {
-            let parts: Vec<&str> = cpu_line.split_whitespace().collect();
-            if parts.len() >= 8 {
-                let idle: u64 = parts[4].parse().unwrap_or(0);
-                let total: u64 = parts[1..8].iter()
-                    .map(|s| s.parse::<u64>().unwrap_or(0))
-                    .sum();
-                if total > 0 {
-                    let usage = 100 - (idle * 100 / total);
-                    return format!("{}%", usage);
-                }
-            }
-        }
-    }
-    "?".to_string()
-}
-
-fn get_network_info() -> String {
-    if let Ok(output) = Command::new("ip").args(&["route", "show", "default"]).output() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        if let Some(line) = stdout.lines().next() {
-            if let Some(interface) = line.split_whitespace().nth(4) {
-                if let Ok(ip_output) = Command::new("ip").args(&["addr", "show", interface]).output() {
-                    let ip_stdout = String::from_utf8_lossy(&ip_output.stdout);
-                    for ip_line in ip_stdout.lines() {
-                        if ip_line.contains("inet ") && !ip_line.contains("127.0.0.1") {
-                            if let Some(ip) = ip_line.split_whitespace().nth(1) {
-                                if let Some(clean_ip) = ip.split('/').next() {
-                                    return format!("{} ({})", interface, clean_ip);
-                                }
-                            }
-                        }
-                    }
-                }
-                return format!("{}", interface);
-            }
-        }
-    }
-    "Нет соединения".to_string()
-}
-
-fn get_locale_info() -> String {
-    env::var("LANG").unwrap_or_else(|_| "?".to_string())
-}
 
 fn get_string_length(s: &str) -> usize {
     let mut result = 0;
